+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.3.0
+
+New shiny API!
+
+### Added
+* Registration methods on container level for root objects.
+* New chained method ```namespace```.
+* Module name validation (doesn't allow to register name with separators)
+
 ## 0.2.0
 
 ### Added